feat: write-push replication layer
Adds a implementation of DmlSink that pushes all writes that pass
through it to an upstream ingester.

This executes in-line with the ingester's normal write path, waiting for
an ACK from the replica before returning an ACK to the user. This is a
naive implementation, with no regard for retries, back-pressure,
batching, etc.