feat: Add a new hash column on the partition table
This will hold the deterministic ID for partitions. Until we backfill
all existing partitions, this is optional/nullable. The row ID still
exists and is used as the foreign key in the parquet_file and
skipped_compaction tables. The hash_id has a unique index so that we can
look up records based on it (if it's available).