fix(iox, gc): delete files from table based on file age
This PR proposes to have the parquet file garbage collector remove files
from the parquet_file table based on the to_delete column being non-null
and the file's create time being older than the soft delete cutoff.

Because to_delete is filled with the timestamp of when the garbage
collector (or compaction) marked the file as soft deleted, it is subject
to the availability of the garbage collector running. If the GC is
unavailable for a significant amount of time (approaching a day or
more), the parquet files will have a to_delete timestamp of when the GC
begins running again (or catches up if it is behind). This further
delays removing files from the parquet table if only to_delete is used.

We know that if to_delete is not null, the file is soft deleted. We know
that the created_at time is immutable so we can use the combination to
determine that a file is available to be deleted from the table.

* helps with influxdata/idpe#17708