fix(compaction,gc): allow a backdating the deleted at in compactions
This CS provides a means to configure a duration to use to back date the
"deleted at" timestamp for parquet files that have been compacted and
are now soft-deleted. This value is in the `to_delete` column of
`parquet_table`. Since we have the newly compacted files, there is
little need to keep the previous version of the data (in the older
files), but various services may be using the compacted soft deleted
files still in active queries or caches.

In combination with the cutoff the garbage collector uses to remove soft
deleted files from the catalog, you can set the adjustment in this CS to
provide for example 20 days of time before the GC will delete from the
catalog the soft deleted files from compaction.

* helps influxdata/idpe#17708