chore: prepare for schema timezone change (#8839)
* chore: prepare for schema timezone change

There are a number of changes required to support setting a timezone
for timestamps in the schema. The changes are mainly so that tests
respect the schema and create correct data.

The schema has not been changed in this commit so there is no change
to user-observable behaviour. These changes were found by changing
the schama timezone fixing all the broken bahaviour and then changing
the timezone back.

* fix: format